Updated module augmentation for upload package.
