+++ packages/ckeditor5-upload/src/adapters/base64uploadadapter.ts
@@ -101,9 +101,3 @@ class Adapter implements UploadAdapter {
 		this.reader!.abort();
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ Base64UploadAdapter.pluginName ]: Base64UploadAdapter;
-	}
-}
