+++ packages/ckeditor5-upload/src/filerepository.ts
@@ -663,9 +663,3 @@ type FilePromiseWrapper = {
 	 */
 	isFulfilled: boolean;
 };
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ FileRepository.pluginName ]: FileRepository;
-	}
-}
