+++ packages/ckeditor5-upload/src/index.ts
@@ -16,3 +16,4 @@ export {
 export { default as FileDialogButtonView } from './ui/filedialogbuttonview';
 export { default as Base64UploadAdapter } from './adapters/base64uploadadapter';
 export { default as SimpleUploadAdapter } from './adapters/simpleuploadadapter';
+export type { SimpleUploadConfig } from './uploadconfig';
